+++ app/Http/Controllers/RoleController.php
@@ -7,6 +7,7 @@ use Spatie\Permission\Models\Permission;
 use Illuminate\Http\Request;
 use App\Http\Requests\RoleStoreRequest;
 use App\Http\Requests\RoleUpdateRequest;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('role_index'), 403);
         $roles = Role::paginate(5);
         return view('roles.index', compact('roles'));
     }
@@ -28,6 +30,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('role_create'), 403);
         $permissions = Permission::all()->pluck('name','id');
         //dd($permissions);
         return view('roles.create', compact('permissions'));
@@ -59,6 +62,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        abort_if(Gate::denies('role_show'), 403);
+
         $role->load('permissions');
         return view('roles.show', compact('role'));
     }
@@ -71,6 +76,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        abort_if(Gate::denies('role_edit'), 403);
+
         $permissions = Permission::all()->pluck('name', 'id');
         //cargamos los permisos relacionados ataves de la relacion muchos a muchos
         $role->load('permissions');
@@ -104,6 +111,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        abort_if(Gate::denies('role_destroy'), 403);
+
         $role->delete();
 
         return redirect()->route('role.index')
